+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "/home/hlrrr/miniconda3/envs/kkakka/bin/python"
+}
